Added access control

+++ app/Http/Controllers/PostsController.php
@@ -13,6 +13,20 @@ use Illuminate\Support\Facades\Auth;
 
 class PostsController extends Controller
 {
+
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+
+    
     /**
      * Display a listing of the resource.
      *
@@ -86,6 +100,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
+
+        if(Auth::user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized User');
+        }
         return view('posts.edit')->with('post',$post);
 
         
@@ -117,6 +135,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post= Post::find($id);
+        if(Auth::user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized User');
+        }
         $post->delete();
         return redirect('/dashboard')->with('success','Post Deleted');
     }
